requirements.txt
# API Server
fastapi==0.94.1
uvicorn[standard]==0.20.0
pywin32; sys_platform == "win32"
python-socketio>=5
aiohttp
tqdm
packaging

# Pre-commit
pre-commit

# TensorRT Diffusion
accelerate
colored
cuda-python
diffusers==0.14.0
ftfy
matplotlib
onnx==1.13.1
onnxruntime==1.14.1
--extra-index-url https://pypi.ngc.nvidia.com
onnx-graphsurgeon==0.3.26
polygraphy==0.44.2
scipy
transformers==4.26.1